refactor: presentation 및 데이터, 네트워크 모듈 분리
