+++ Near/app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.Near"
         tools:targetApi="31">
         <activity
-            android:name=".feature.main.MainActivity"
+            android:name=".presentation.feature.main.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Near">
