+++ Near/app/src/main/java/com/alarmy/near/data/di/RepositoryModule.kt
@@ -1,0 +1,17 @@
+package com.alarmy.near.data.di
+
+import com.alarmy.near.data.repository.ExampleRepository
+import com.alarmy.near.data.repository.ExampleRepositoryImpl
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+interface RepositoryModule {
+    @Binds
+    @Singleton
+    abstract fun bindExampleRepository(exampleRepositoryImpl: ExampleRepositoryImpl): ExampleRepository
+}
