+++ Near/app/src/main/java/com/alarmy/near/data/repository/ExampleRepositoryImpl.kt
@@ -1,0 +1,9 @@
+package com.alarmy.near.data.repository
+
+import javax.inject.Inject
+
+class ExampleRepositoryImpl
+    @Inject
+    constructor() : ExampleRepository {
+        override fun getExampleData(): String = "Hello from Repository"
+    }
