+++ Near/app/src/main/java/com/alarmy/near/network/di/AppModule.kt
@@ -1,4 +1,4 @@
-package com.alarmy.near.di
+package com.alarmy.near.network.di
 
 // @Module
 // @InstallIn(SingletonComponent::class)
