+++ Near/app/src/main/java/com/alarmy/near/presentation/feature/home/navigation/HomeNavigation.kt
@@ -1,10 +1,10 @@
-package com.alarmy.near.feature.home.navigation
+package com.alarmy.near.presentation.feature.home.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
-import com.alarmy.near.feature.home.HomeRoute
+import com.alarmy.near.presentation.feature.home.HomeRoute
 import kotlinx.serialization.Serializable
 
 @Serializable
