+++ Near/app/src/main/java/com/alarmy/near/presentation/feature/main/MainActivity.kt
@@ -1,10 +1,10 @@
-package com.alarmy.near.feature.main
+package com.alarmy.near.presentation.feature.main
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import com.alarmy.near.ui.theme.NearTheme
+import com.alarmy.near.presentation.ui.theme.NearTheme
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
