+++ Near/app/src/main/java/com/alarmy/near/presentation/feature/main/NearNavHost.kt
@@ -1,11 +1,11 @@
-package com.alarmy.near.feature.main
+package com.alarmy.near.presentation.feature.main
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
-import com.alarmy.near.feature.home.navigation.RouteHome
-import com.alarmy.near.feature.home.navigation.homeNavGraph
+import com.alarmy.near.presentation.feature.home.navigation.RouteHome
+import com.alarmy.near.presentation.feature.home.navigation.homeNavGraph
 
 @Composable
 internal fun NearNavHost(
