+++ Near/app/src/main/java/com/alarmy/near/presentation/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.alarmy.near.ui.theme
+package com.alarmy.near.presentation.ui.theme
 
 import androidx.compose.ui.graphics.Color
