--- Near/app/src/main/java/com/alarmy/near/repository/ExampleRepositoryImpl.kt
@@ -1,5 +1,0 @@
-package com.alarmy.near.repository
-
-class ExampleRepositoryImpl : ExampleRepository {
-    override fun getExampleData(): String = "Hello from Repository"
-}
